requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
backtrader>=1.9.78.123
matplotlib>=3.7.0
pandas-ta
streamlit
plotly
scikit-learn>=1.2.0
torch>=2.0.0
gymnasium>=0.28.1
stable-baselines3>=2.0.0
shimmy>=0.2.1
feedparser>=6.0.10
vaderSentiment>=3.3.2
web3>=6.0.0
python-telegram-bot>=20.0
psutil>=5.9.0
extra-streamlit-components>=0.1.81
nixtla>=0.3.0
# chronos-forecasting  # Commented out to avoid heavy install issues during dev, can be enabled for prod
torch>=2.0.0
